In process of separating systems and components
- Physics
- Collision
- Graphics
- RTTR

+++ Merge Attempt/maki/SushiMi/Engine/Components/Physics/collision.cpp
@@ -63,9 +63,9 @@ namespace Core
 		Step 5: Otherwise the rectangles intersect
 
 		*/
-		float tFirst = 0.f;
-		float tLast = g_dt;
-		float dFirst = 0.f, dLast = 0.f;
+		double tFirst = 0.f;
+		double tLast = g_dt;
+		double dFirst = 0.f, dLast = 0.f;
 		if (aabb1.max.y < aabb2.min.y || aabb1.max.x < aabb2.min.x || aabb1.min.y > aabb2.max.y || aabb1.min.x > aabb2.max.x) //if no overlap, continue
 		{
 			return 0;
